requirements.txt
nose
requests
vcrpy
